fix: move scene setup build hooks into editor assembly
